+++ HotItemsAnalysis/hot_items_with_sql.py
@@ -12,7 +12,7 @@ from pyflink.common.watermark_strategy import TimestampAssigner, WatermarkStrate
 
 def split_data(row):
     line = row.split(',')
-    user_id, item_id, category_id, behavior, timestamp = int(line[0]), int(line[1]), int(line[2]), line[3], float(line[4])
+    user_id, item_id, category_id, behavior, timestamp = int(line[0]), int(line[1]), int(line[2]), line[3], int(line[4])
     yield Row(user_id, item_id, category_id, behavior, timestamp)
