+++ LoginFailDetect/loginfail.py
@@ -11,7 +11,7 @@ from pyflink.util.java_utils import get_j_env_configuration
 
 def split_input(row):
     line = row.split(',')
-    user_id, ip, event_type, timestamp = int(line[0]), line[1], line[2], float(line[3])
+    user_id, ip, event_type, timestamp = int(line[0]), line[1], line[2], int(line[3])
     yield user_id, ip, event_type, timestamp
